% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbowPoint.R
\name{elbowPoint}
\alias{elbowPoint}
\title{Elbow (or Knee) Point Detection Function
#'
This is adapted from an R script written by L. Pereira based on
a solution by Jonas on stackoverflow}
\usage{
elbowPoint(x, y)
}
\arguments{
\item{x}{Date range in julian date, origin = '2019-12-31'}

\item{y}{Rate of change, k in y = Ae^(kx)}
}
\value{
Date of elbow changepoint in Julian and Date format
}
\description{
This function detects the change point of a curve with
an elbow (or knee) shape
}
\examples{
elbowPoint(x, y)
}
