% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_breakpts.R
\name{intervention_breakpts}
\alias{intervention_breakpts}
\title{Intervention Breakpoint Deetection Function}
\usage{
intervention_breakpts(x, y)
}
\arguments{
\item{x}{Date range in julian date, origin = '2019-12-31'}

\item{y}{Rate of change, k in y = Ae^(kx)}
}
\value{
Date of elbow changepoint in Julian and Date format
}
\description{
This function detects breakpoints where the slope of
rates of cases change
}
\details{
This is adapted from an R script written by Markus Loew on RPubs.
breakpoints() in struccchange package by Achim Zeileis was used
to estimate psi param in segmentation function
}
\examples{
intervention_breakpts(x, y)
}
